requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure Functions
azure-functions==1.17.0

# Azure OpenAI Integration
openai>=1.0.0
azure-identity>=1.14.1
azure-keyvault-secrets>=4.7.0

# Database (Optional - for future use)
# sqlalchemy==2.0.23
# asyncpg==0.29.0
# alembic==1.12.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
python-dotenv==1.0.0

# Logging and Monitoring
python-json-logger==2.0.7

# Configuration
pyyaml>=6.0.1

# Language Detection
langdetect==1.0.9

# Rate Limiting (Optional)
# slowapi==0.1.9
